requirements.txt
fastapi==0.115.8
faster_whisper==1.1.1
keyboard==0.13.5
langchain_core==0.3.33
langchain_google_vertexai==2.0.13
langchain_groq==0.2.4
langchain_ollama==0.2.3
langchain_openai==0.3.3
langserve==0.3.1
nest_asyncio==1.6.0
openai==1.61.0
Pillow==11.1.0
protobuf==5.29.3
pyaudio==0.2.14
python-dotenv==1.0.1
pyttsx3==2.98
uvicorn==0.34.0
